Normal mapping!
Normal mapping is now working! All the light sources should now correctly light up a model with a normal map, as there is the one main lighting shader there is a default normal map that can be used for objects without one.
The pre set models are not set up correctly with the tangents and bi tangents
Texture has an unload function, this is now called via the resource manager
Debug stuff set up for glsl
Made one of the point lights move back and forth