Animation Bone Cleanup
Bones now assume that the amount of keys for position, rotation and scales will be the same amount.
The key structs (such as KeyPosition, KeyRotation and KeyScale) have been replaced with a single struct for a whole key, which contains a position, rotation, scale and timestamp.
Bone interpolate and get index functions have also been changed to be for a whole key at once