Screen drawn with framebuffer
Texture has a setWidthHeight function, this calls load on the texture after setting the width and height to the given values.
A fullscreen framebuffer is now used to draw the scene.
Multisampling has been disabled, needs to be implemented for the framebuffer.
Framebuffer now can use a render buffer object.
Framebuffer refresh textures has been renamed to Load, it also now re creates the RBO if it is being used.
Framebuffer now has a setWidthHeight function that sets the width and height and re loads the framebuffer